Translate command refactoring
